838: working on N=40,2800 test cases, but not on N=1million
Using a more general way to replace some pAs with pBs, allowing for a
list of length n of pAs that all have the same list of pB cofactors,
to be replaced with that list of pB cofactors if the
pBa if the prod(cofactors) < prod(pAsList). But this doesn't solve the
problem, sadly.

Also added the N=1million "test" case, modified logging (including
forcing output stream flushing for logDebug)